requirements.txt
# Core
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0

# Image & Video Processing
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0

# Data Download & Management
roboflow>=1.1.0
requests>=2.31.0
tqdm>=4.66.0
pyyaml>=6.0

# Alert & UI
tkinter-tooltip>=2.1.0

# Visualization & Logging
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.7
